Moving SNES delegates into SNESolver
