+++ include/godzilla/SNESolver.h
@@ -9,6 +9,84 @@
 
 namespace godzilla {
 
+namespace internal {
+
+/// Abstract "method" for calling monitor
+struct SNESMonitorMethodAbstract {
+    virtual ~SNESMonitorMethodAbstract() = default;
+    virtual ErrorCode invoke(Int it, Real rnorm) = 0;
+};
+
+template <typename T>
+struct SNESMonitorMethod : public SNESMonitorMethodAbstract {
+    SNESMonitorMethod(T * instance, ErrorCode (T::*monitor)(Int, Real)) :
+        instance(instance),
+        monitor(monitor)
+    {
+    }
+
+    ErrorCode
+    invoke(Int it, Real rnorm) override
+    {
+        return ((*this->instance).*monitor)(it, rnorm);
+    }
+
+private:
+    T * instance;
+    ErrorCode (T::*monitor)(Int, Real);
+};
+
+/// Abstract "method" for calling compute residual
+struct SNESComputeResidualMethodAbstract {
+    virtual ErrorCode invoke(const Vector & x, Vector & f) = 0;
+};
+
+template <typename T>
+struct SNESComputeResidualMethod : public SNESComputeResidualMethodAbstract {
+    SNESComputeResidualMethod(T * instance, ErrorCode (T::*method)(const Vector &, Vector &)) :
+        instance(instance),
+        method(method)
+    {
+    }
+
+    ErrorCode
+    invoke(const Vector & x, Vector & f) override
+    {
+        return ((*this->instance).*method)(x, f);
+    }
+
+private:
+    T * instance;
+    ErrorCode (T::*method)(const Vector & x, Vector & f);
+};
+
+/// Abstract "method" for calling compute Jacobian
+struct SNESComputeJacobianMethodAbstract {
+    virtual ErrorCode invoke(const Vector & x, Matrix & J, Matrix & Jp) = 0;
+};
+
+template <typename T>
+struct SNESComputeJacobianMethod : public SNESComputeJacobianMethodAbstract {
+    SNESComputeJacobianMethod(T * instance,
+                              ErrorCode (T::*method)(const Vector &, Matrix &, Matrix &)) :
+        instance(instance),
+        method(method)
+    {
+    }
+
+    ErrorCode
+    invoke(const Vector & x, Matrix & J, Matrix & Jp) override
+    {
+        return ((*this->instance).*method)(x, J, Jp);
+    }
+
+private:
+    T * instance;
+    ErrorCode (T::*method)(const Vector &, Matrix &, Matrix &);
+};
+
+} // namespace internal
+
 /// Wrapper around SNES
 class SNESolver {
 public:
